Added code to update a user's details.
This code just does what it says. It will update the user details based on the update object. So, we have to set the details in the update object ourselves. It will depend on what the user has entered to change. The code makes use of the hashPassword function from users.js to hash whatever password was entered by the user. The rest is straightforward: It makes use of the updateOne built in function and the $set notation to update details in mongodb collection users.